Revert requirements.txt to proper format.
requirements.txt is a structured file format used by pip (and pip3) to
provide an interface for describing the packages (and versions thereof)
needed for an application. To verify a requirements file, you can use
the command 'pip3 install -r requirements.txt' where the '-r' flag
indicates you're providing a file with the aforementioned format.